third_party/blink/web_tests/svg/dynamic-updates/SVGMaskElement-dom-y-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'y' attribute of the SVGMaskElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS maskElement.getAttribute('y') is "-100%"
PASS maskElement.getAttribute('y') is "-10%"
PASS successfullyParsed is true

TEST COMPLETE
